moved game speed to memory
