+++ nqp/scenes/world/scene.py
@@ -47,7 +47,7 @@ class WorldScene(Scene):
 
     def update(self, delta_time: float):
         # get the modified delta time
-        mod_delta_time = self.model.game_speed * delta_time
+        mod_delta_time = self._game.memory.game_speed * delta_time
 
         # update the data
         self.model.update(mod_delta_time)
